input error 문구추가 함수 모듈화
